ev-charging-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.9
pydantic==2.10.5
httpx==0.27.2
python-dotenv==1.0.1
